refactor: reduce duplicated code between math-picker and canvas-picker
