+++ packages/vrender-kits/src/picker/contributions/math-picker/circle-picker.ts
@@ -1,9 +1,11 @@
-import { inject, injectable, CircleRender } from '@visactor/vrender-core';
+import { inject, injectable, CircleRender, CIRCLE_NUMBER_TYPE } from '@visactor/vrender-core';
 import type { IGraphicPicker, IGraphicRender } from '@visactor/vrender-core';
-import { CirclePickerBase } from '../common/circle-picker-base';
+import { PickerBase } from '../common/base';
 
 @injectable()
-export class DefaultMathCirclePicker extends CirclePickerBase implements IGraphicPicker {
+export class DefaultMathCirclePicker extends PickerBase implements IGraphicPicker {
+  type: string = 'circle';
+  numberType: number = CIRCLE_NUMBER_TYPE;
   constructor(@inject(CircleRender) public readonly canvasRenderer: IGraphicRender) {
     super();
   }
